+++ .ci/install_linux.sh
@@ -1,3 +1,6 @@
+#!/usr/bin/env bash
+set -ex
+
 sudo apt-add-repository --yes ppa:libccd-debs/ppa
 sudo apt-add-repository --yes ppa:fcl-debs/ppa
 sudo apt-add-repository --yes ppa:dartsim/ppa
@@ -6,7 +9,9 @@ sudo apt-get -qq update
 APT='
 cmake
 libassimp-dev
-libboost-all-dev
+libboost-filesystem-dev
+libboost-regex-dev
+libboost-system-dev
 libccd-dev
 libeigen3-dev
 libfcl-dev