+++ dart/common/Memory.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2017, The DART development contributors
+ * Copyright (c) 2011-2018, The DART development contributors
  * All rights reserved.
  *
  * The list of contributors can be found at:
@@ -38,6 +38,7 @@
 
 #include "dart/config.hpp"
 #include "dart/common/Deprecated.hpp"
+#include "dart/common/detail/AlignedAllocator.hpp"
 
 namespace dart {
 namespace common {