+++ dart/dynamics/ReferentialSkeleton.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2017, The DART development contributors
+ * Copyright (c) 2011-2018, The DART development contributors
  * All rights reserved.
  *
  * The list of contributors can be found at:
@@ -34,6 +34,7 @@
 #define DART_DYNAMICS_REFERENTIALSKELETON_HPP_
 
 #include <unordered_map>
+#include <unordered_set>
 
 #include "dart/dynamics/MetaSkeleton.hpp"
 #include "dart/dynamics/SmartPointer.hpp"
@@ -54,6 +55,9 @@ public:
   /// Default destructor
   virtual ~ReferentialSkeleton() = default;
 
+  // Documentation inherited
+  std::unique_ptr<common::LockableReference> getLockableReference() const override;
+
   //----------------------------------------------------------------------------
   /// \{ \name Name
   //----------------------------------------------------------------------------
@@ -69,6 +73,13 @@ public:
   //----------------------------------------------------------------------------
   /// \{ \name Structural Properties
   //----------------------------------------------------------------------------
+
+  /// Returns number of skeletons associated with this ReferentialSkeleton.
+  std::size_t getNumSkeletons() const;
+
+  /// Returns whether this ReferentialSkeleton contains any BodyNode or Joint
+  /// from \c skel.
+  bool hasSkeleton(const Skeleton* skel) const;
 
   // Documentation inherited
   std::size_t getNumBodyNodes() const override;
@@ -115,6 +126,9 @@ public:
       const std::string& name) const override;
 
   // Documentation inherited
+  bool hasBodyNode(const BodyNode* bodyNode) const override;
+
+  // Documentation inherited
   std::size_t getIndexOf(const BodyNode* _bn, bool _warning=true) const override;
 
   // Documentation inherited
@@ -159,6 +173,9 @@ public:
   /// \note ReferentialSkeleton can contain multiple Joints with the same
   /// name when ReferentialSkeleton contains Joints from multiple Skeletons.
   std::vector<const Joint*> getJoints(const std::string& name) const override;
+
+  // Documentation inherited
+  bool hasJoint(const Joint* joint) const override;
 
   // Documentation inherited
   std::size_t getIndexOf(const Joint* _joint, bool _warning=true) const override;
@@ -454,6 +471,14 @@ protected:
   /// Name of this ReferentialSkeleton
   std::string mName;
 
+  /// Skeletons that this ReferentialSkeleton contains any BodyNode or Joint
+  /// from the Skeletons.
+  std::unordered_set<const Skeleton*> mSkeletons;
+
+  /// Mutexes of the skeletons. The mutexes are sorted in order of memory
+  /// addresses.
+  std::set<std::mutex*> mSkeletonMutexes;
+
   /// BodyNodes that this ReferentialSkeleton references. These hold strong
   /// references to ensure that the BodyNodes do not disappear
   std::vector<BodyNodePtr> mBodyNodes;
@@ -509,6 +534,15 @@ protected:
 
   /// Cache for constraint force vector
   mutable Eigen::VectorXd mFc;
+
+private:
+  /// Add a Skeleton to this ReferentialSkeleton, ignoring its Joint and
+  /// DegreesOfFreedom. This can only be used by this class.
+  void registerSkeleton(const Skeleton* skel);
+
+  /// Removes a Skeleton from this ReferentialSkeleton. This can only be used by
+  /// this class.
+  void unregisterSkeleton(const Skeleton* skel);
 };
 
 } // namespace dynamics